I Got the Box Moving
The box is now moving…. in all directions (using the directional keys
on the keyboard) added some collision detection to make sure it didn't
go below the "grass" Still a work in progress. next is to get some side
and top barriers so that I can create some collision detection to keep
the box..…n a box.